[BATLLE-2] Refactor BattlefieldStorageVisual into Data, Component, and Visual
