+++ .idea/.idea.File_Commander/.idea/workspace.xml
@@ -92,7 +92,7 @@
       <workItem from="1725319130277" duration="31726000" />
       <workItem from="1727650723952" duration="24174000" />
       <workItem from="1728099179833" duration="1800000" />
-      <workItem from="1728250890335" duration="11177000" />
+      <workItem from="1728250890335" duration="11932000" />
     </task>
     <servers />
   </component>
